Implement Extended Enconding, but got problem worst at correctness

+++ core/sudoku2dimacs.py
@@ -3,7 +3,7 @@ import sudoku
 from pathlib import Path
 
 def calculateClause(nCell):
-    return nCell*nCell + (nCell*nCell * (nCell * (nCell - 1)/2)) * 3
+    return int(nCell*nCell + (nCell*nCell * (nCell * (nCell - 1)/2)) * 3)
 
 def main(sudokuProblem):
